// src/app/features/links/interfaces/response-update-link.interface.ts
export interface ResponseUpdateLinkInterface {
  linkUpdated: LinkUpdated;
  message: string;
}

export interface LinkUpdated {
  id: string;
  originalUrl: string;
  shortCode: string;
  customAlias: string;
  title: string;
  description: string;
  password: string;
  expiresAt: Date | null;
  isActive: boolean;
  isPublic: boolean;
  category: string;
  userId: string;
  created_at: Date;
  updated_at: Date;
}